Rename Users to User
Update UserService methods: add private
