+++ transcendence-app/src/oauth42/oauth42.module.ts
@@ -1,11 +1,11 @@
 import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
-import { UsersModule } from '../users/users.module';
+import { UserModule } from '../user/user.module';
 import { OAuth42Controller } from './oauth42.controller';
 import { OAuth42Strategy } from './oauth42.strategy';
 
 @Module({
-  imports: [HttpModule, UsersModule],
+  imports: [HttpModule, UserModule],
   controllers: [OAuth42Controller],
   providers: [OAuth42Strategy],
 })
